+++ CHANGELOG.md
@@ -9,3 +9,9 @@ This file is used to list changes made in each version of vagrant.
 ## 1.3.5:
 
 * Updated default Vagrant version to 1.3.5
+
+## 1.4.0:
+
+* Updated default Vagrant version to 1.4.3
+* Updated to work with new download endpoint(http://www.vagrantup.com/downloads.html)
+* Add optional version attribute for vagrant plugins